Version 6

+++ CalculatorWpfApplication/CalculatorWpfApplication/MathOperations/Logic/MathExpressionsLogic.cs
@@ -4,7 +4,7 @@ namespace CalculatorWpfApplication
 {
     public class MathExpressionsLogic
     {
-        public static int Addition(int firstElement, int result) {return (result+=firstElement);}
+        public static double Addition(double firstElement, double result) {return (result+=firstElement);}
 
     }
 }
